Backend/requirements.txt
fastapi[standard]>=0.115.0
uvicorn[standard]>=0.30.0
gunicorn>=21.2.0
pymongo[srv]>=4.8.0
beautifulsoup4>=4.12.3
requests>=2.32.3
debugpy>=1.8.0
pydantic>=2.12.5
uv>=0.4.0
pytest>=8.3.3
ruff>=0.7.1
pydantic-ai
python-dotenv
